gigs-in-town/pyproject.toml
[project]
name = "gigs-in-town"
version = "0.1.0"
description = "Process ICS files to deliver statistics on concert events"
readme = "README.md"
authors = [
    { name = "Italo Silva", email = "italo@maleldil.com" }
]
requires-python = ">=3.12"
dependencies = [
    "icalendar>=6.1.0",
    "pydantic>=2.10.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "typer>=0.13.1",
]
license = {text = "GPL-3.0-or-later"}

[project.scripts]
gigs-in-town = "gigs_in_town.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.389",
    "ruff>=0.8",
]

[tool.ruff.lint.isort]
known-first-party = ["gigs_in_town"]

[tool.ruff.lint]
select = [
    "E",        # Code style
    "W",        # Code style
    "F",        # Pyflakes
    "I",        # Imports
    "ISC",      # Implicit string concat
    "N",        # PEP-8 naming
    "PL",       # Pylint
    "RUF",      # Ruff-specific
    "SIM",      # Simplify
    "TRY",      # Exceptions
    "UP",       # Upgrade to new constructs
    "ERA",      # Commented-out code
    "D",        # Documentation
    "FURB",     # Modernisation
    "PTH",      # Pathlib usage
    "RET",      # Unnecessary elses
    "G",        # Logging format
    "DTZ",      # Timezones
    "B",        # Likely bugs
    "C",        # Comprehensions
    "ASYNC",    # Async pitfalls
    "EXE",      # Executable (shebang)S
    "ANN",      # Type annotations
    "PYI",      # Appropriate type hints for specifc cases
    "PERF",     # Performance anti-patterns
]
ignore = [
    "PLR0911",  # Too many return statements
    "PLR0912",  # Too many branches
    "PLR0913",  # Too many arguments
    "PLR0915",  # Too many statements
    "PLR2004",  # Magic value used in comparison
    "E501",     # Line too long (let ruff format take care of this)
    "TRY003",   # Long messages outside exception class
    "SIM108",   # Use ternary operation instead of if-else block
    "ISC001",   # Incompatible with the formatter
    "D202",     # Empty lines between function docstring and body
    "D1",       # Undocumented items
    "PTH123",   # `open` instead of `Path.open`
    "G004",     # f-strings in logging
    "B905",     # `zip` without `strict`
    "C901",     # "too complex" functions
    "ASYNC110", # 'asyncio.sleep' in loop
    "PERF401",  # Manual list comprehensions
    "ANN401",   # Any
    "PLC0206",  # Dictionary iteration without items
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
# Ruff also handles this, and it needs to be enabled there for fixAll to fix it,
# so we disable it here to avoid duplicate errors
reportUnusedImport = false
typeCheckingMode = "strict"

# Allow unit tests to import private things
executionEnvironments = [
  { root = "tests", reportPrivateUsage = false }
]
